[fbcnms-ui][fbcnms-alarms] fix notistack upgrade
