+++ fbcnms-packages/fbcnms-alarms/components/alertmanager/Receivers/Receivers.js
@@ -23,7 +23,7 @@ import TableAddButton from '../../table/TableAddButton';
 import {makeStyles} from '@material-ui/styles';
 import {useAlarmContext} from '../../AlarmContext';
 import {useNetworkId} from '../../../components/hooks';
-import {useSnackbars} from '../../../hooks/useSnackbar';
+import {useSnackbars} from '@fbcnms/ui/hooks/useSnackbar';
 
 import type {AlertReceiver} from '../../AlarmAPIType';
