+++ fbcnms-packages/fbcnms-alarms/components/alertmanager/Receivers/__tests__/Receivers-test.js
@@ -21,10 +21,10 @@ import {alarmTestUtil, useMagmaAPIMock} from '../../../../test/testHelpers';
 
 const enqueueSnackbarMock = jest.fn();
 jest
-  .spyOn(require('../../../../hooks/useSnackbar'), 'useEnqueueSnackbar')
+  .spyOn(require('@fbcnms/ui/hooks/useSnackbar'), 'useEnqueueSnackbar')
   .mockReturnValue(enqueueSnackbarMock);
 jest
-  .spyOn(require('../../../../hooks/useRouter'), 'default')
+  .spyOn(require('@fbcnms/ui/hooks/useRouter'), 'default')
   .mockReturnValue({match: {params: {networkId: 'test'}}});
 
 const {AlarmsWrapper} = alarmTestUtil();
