+++ fbcnms-packages/fbcnms-alarms/components/alertmanager/Suppressions.js
@@ -17,7 +17,7 @@ import TableActionDialog from '../table/TableActionDialog';
 import {makeStyles} from '@material-ui/styles';
 import {useAlarmContext} from '../AlarmContext';
 import {useNetworkId} from '../../components/hooks';
-import {useSnackbars} from '../../hooks/useSnackbar';
+import {useSnackbars} from '@fbcnms/ui/hooks/useSnackbar';
 
 import {useState} from 'react';
