+++ fbcnms-packages/fbcnms-alarms/components/rules/PrometheusEditor/__tests__/PrometheusEditor-test.js
@@ -17,17 +17,17 @@ import {render} from '@testing-library/react';
 import type {AlertConfig} from '../../../AlarmAPIType';
 import type {GenericRule} from '../../RuleInterface';
 
-jest.mock('../../../../hooks/useSnackbar');
-jest.mock('../../../../hooks/useRouter');
+jest.mock('@fbcnms/ui/hooks/useSnackbar');
+jest.mock('@fbcnms/ui/hooks/useRouter');
 
 const {AlarmsWrapper} = alarmTestUtil();
 
 const enqueueSnackbarMock = jest.fn();
 jest
-  .spyOn(require('../../../../hooks/useSnackbar'), 'useEnqueueSnackbar')
+  .spyOn(require('@fbcnms/ui/hooks/useSnackbar'), 'useEnqueueSnackbar')
   .mockReturnValue(enqueueSnackbarMock);
 jest
-  .spyOn(require('../../../../hooks/useRouter'), 'default')
+  .spyOn(require('@fbcnms/ui/hooks/useRouter'), 'default')
   .mockReturnValue({match: {params: {networkId: 'test'}}});
 
 // TextField select is difficult to test so replace it with an Input
