+++ fbcnms-packages/fbcnms-ui/hooks/useSnackbar.js
@@ -30,7 +30,7 @@ export default function useSnackbar(
     if (show) {
       const config: AllowedConfig = JSON.parse(stringConfig);
       const k = enqueueSnackbar(message, {
-        children: key => (
+        content: key => (
           <SnackbarItem
             id={key}
             message={message}
@@ -61,9 +61,9 @@ export default function useSnackbar(
 export function useEnqueueSnackbar() {
   const {enqueueSnackbar} = useNotistackSnackbar();
   return useCallback(
-    (message: string, config: Object) =>
+    (message: string, config: EnqueueSnackbarOptions) =>
       enqueueSnackbar(message, {
-        children: key => (
+        content: key => (
           <SnackbarItem
             id={key}
             message={message}
@@ -74,4 +74,44 @@ export function useEnqueueSnackbar() {
       }),
     [enqueueSnackbar],
   );
+}
+
+export function useSnackbars() {
+  const enqueueSnackbar = useEnqueueSnackbar();
+
+  const successSnackbar = React.useCallback(
+    (message: string) =>
+      enqueueSnackbar(message, {
+        variant: 'success',
+      }),
+    [enqueueSnackbar],
+  );
+
+  const errorSnackbar = React.useCallback(
+    (message: string) => {
+      enqueueSnackbar(message, {
+        variant: 'error',
+      });
+    },
+    [enqueueSnackbar],
+  );
+
+  const warningSnackbar = React.useCallback(
+    (message: string) =>
+      enqueueSnackbar(message, {
+        variant: 'warning',
+      }),
+    [enqueueSnackbar],
+  );
+
+  const result = React.useMemo(
+    () => ({
+      success: successSnackbar,
+      error: errorSnackbar,
+      warning: warningSnackbar,
+    }),
+    [errorSnackbar, successSnackbar, warningSnackbar],
+  );
+
+  return result;
 }
