+++ fbcnms-packages/fbcnms-ui/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@fbcnms/ui",
-  "version": "0.1.10",
+  "version": "0.1.11",
   "dependencies": {
     "@material-ui/core": "^4.0.0",
     "@material-ui/icons": "^4.0.0",
